requirements.txt
asgiref==3.7.2
Django==4.2.7
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.14.0
django-compressor==4.4